Create BaseAlphaRouter
