+++ src/routers/alpha-router/functions/best-swap-route.ts
@@ -11,7 +11,7 @@ import { CurrencyAmount } from '../../../util/amounts';
 import { log } from '../../../util/log';
 import { metric, MetricLoggerUnit } from '../../../util/metric';
 import { routeAmountsToString, routeToString } from '../../../util/routes';
-import { AlphaRouterConfig } from '../alpha-router';
+import { AlphaRouterConfig } from '../base-alpha-router';
 import { IGasModel, L1ToL2GasCosts, usdGasTokensByChain } from '../gas-models';
 
 import { RouteWithValidQuote, V3RouteWithValidQuote, } from './../entities/route-with-valid-quote';
