+++ src/routers/alpha-router/index.ts
@@ -1,3 +1,4 @@
+export * from './base-alpha-router';
 export * from './alpha-router';
 export * from './entities';
 export * from './gas-models';
