+++ src/routers/alpha-router/quoters/base-quoter.ts
@@ -7,7 +7,7 @@ import _ from 'lodash';
 import { ITokenListProvider, ITokenProvider, ITokenValidatorProvider, TokenValidationResult } from '../../../providers';
 import { ChainId, CurrencyAmount, log, poolToString } from '../../../util';
 import { MixedRoute, V2Route, V3Route } from '../../router';
-import { AlphaRouterConfig } from '../alpha-router';
+import { AlphaRouterConfig } from '../base-alpha-router';
 import { RouteWithValidQuote } from '../entities/route-with-valid-quote';
 import { CandidatePoolsBySelectionCriteria } from '../functions/get-candidate-pools';
 import { IGasModel } from '../gas-models';
